+++ noodles-vcf/CHANGELOG.md
@@ -10,6 +10,10 @@
 
     This is meant as a convenience to more easily create a mutable buffer.
 
+  * vcf/record: Implement `TryFrom<(&Header, &str)>`.
+
+    This can be used to parse a raw record.
+
   * vcf/record/ids: Implement `Extend<Id>` and `FromIterator<Id>`.
 
 ### Changed